requirements.txt
# Core ML Libraries
torch==2.1.0
transformers==4.36.2
datasets==2.15.0
accelerate==0.25.0
peft==0.7.1
bitsandbytes==0.41.3

# Evaluation & Monitoring
evaluate==0.4.1
wandb==0.16.1
rouge-score==0.1.2
bert-score==0.3.13  # Changed from 0.9.0

# API & Deployment
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
pyyaml==6.0.1
prometheus-client==0.19.0
redis==5.0.1
httpx==0.25.2